Security Api + DB

+++ java-api/src/main/java/com/db/grad/javaapi/repository/SecurityRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.db.grad.javaapi.model.Security;
 
-public interface SecurityRepository extends JpaRepository<Security, Long> {
+
+public interface SecurityRepository extends JpaRepository<Security, Integer> {
 	
 }
